pyproject.toml
[project]
name = "repo-scout"
version = "0.1.0"
dependencies = [
    "typer[all]>=0.12.0",
    "pydantic-settings>=2.0.0",  # Good for CLI config management
    "pyyaml>=6.0.0",             # For config files
    "rich>=13.0.0",              # For pretty output
    "pydantic>=2.0.0",           # For data models
    "pydantic-core>=2.0.0",      # For data model validation
    "httpx>=0.27.0",             # For GitHub API
    "textual>=6.11.0",           # For TUI output
]
requires-python = ">=3.9"

[project.optional-dependencies]
schema = ["jsonschema>=4.21.0"]

[project.scripts]
# This creates the actual terminal command
scout = "scout.cli.app:app"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"] # pycodestyle, pyflakes, isort, pyupgrade

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src/scout"]

[tool.hatch.build.targets.wheel.force-include]
"src/scout/rules/builtins" = "scout/rules/builtins"
"src/scout/rules/schema.json" = "scout/rules/schema.json"

[build-system]
requires = ["hatchling"]  # or "poetry-core"
build-backend = "hatchling.build"
